+++ config/metable.php
@@ -59,4 +59,22 @@ return [
             ],
         ],
     ],
+
+    /**
+     * Whether to index complex data types (arrays, objects, etc).
+     * If enabled the value will be serialized and the first 255 characters will be indexed.
+     * This allows for using whereMeta*() query scopes on serialized values, but may have
+     * performance and disk usage implications for large data sets.
+     *
+     * If you do not intend to query meta values containing complex data types, you should leave this disabled.
+     */
+    'indexComplexDataTypes' => false,
+
+    /**
+     * Number of bytes to index for strings and complex data types.
+     * This value is used to determine the length of the index column in the database.
+     * Higher values allow for better precision when querying,
+     * but will use more disk space in the database.
+     */
+    'stringValueIndexLength' => 255,
 ];