+++ migrations/2024_04_14_000000_add_meta_search_columns.php
@@ -2,6 +2,7 @@
 
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Schema;
 
 class AddMetaSearchColumns extends Migration
@@ -14,9 +15,13 @@ class AddMetaSearchColumns extends Migration
     public function up()
     {
         Schema::table('meta', function (Blueprint $table) {
-            $table->decimal('numeric_value', 18, 9)->nullable();
-            $table->string('string_value', 255)->nullable();
+            $table->decimal('numeric_value', 36, 16)->nullable();
+            $table->string(
+                'string_value',
+                config('metable.stringValueIndexLength', 255)
+            )->nullable();
             $table->dropIndex(['key', 'metable_type']);
+            $table->dropIndex(['key']);
             $table->index(['key', 'metable_type', 'numeric_value']);
             $table->index(['key', 'metable_type', 'string_value']);
         });
@@ -32,9 +37,10 @@ class AddMetaSearchColumns extends Migration
         Schema::table('meta', function (Blueprint $table) {
             $table->dropIndex(['key', 'metable_type', 'string_value']);
             $table->dropIndex(['key', 'metable_type', 'numeric_value']);
-            $table->index(['metable_type', 'metable_id']);
-            $table->dropColumn('numeric_value');
+            $table->index(['key']);
+            $table->index(['key', 'metable_type']);
             $table->dropColumn('string_value');
+            $table->dropColumn('numeric_value');
         });
     }
 }