+++ src/DataType/FloatHandler.php
@@ -20,12 +20,12 @@ class FloatHandler extends ScalarHandler
         return 'float';
     }
 
-    public function getNumericValue(mixed $value, string $serializedValue): null|int|float
+    public function getNumericValue(mixed $value): null|int|float
     {
         return $value;
     }
 
-    public function getStringValue(mixed $value, string $serializedValue): null|string
+    public function getStringValue(mixed $value): null|string
     {
         return (string) $value;
     }