+++ src/DataType/ModelHandler.php
@@ -57,13 +57,18 @@ class ModelHandler implements HandlerInterface
         return $class::query()->find($id);
     }
 
-    public function getNumericValue(mixed $value, string $serializedValue): null|int|float
+    public function getNumericValue(mixed $value): null|int|float
     {
         return null;
     }
 
-    public function getStringValue(mixed $value, string $serializedValue): null|string
+    public function getStringValue(mixed $value): null|string
     {
-        return $serializedValue;
+        return $this->serializeValue($value);
+    }
+
+    public function isIdempotent(): bool
+    {
+        return true;
     }
 }