+++ src/DataType/NullHandler.php
@@ -20,12 +20,12 @@ class NullHandler extends ScalarHandler
         return 'null';
     }
 
-    public function getNumericValue(mixed $value, string $serializedValue): null|int|float
+    public function getNumericValue(mixed $value): null|int|float
     {
         return null;
     }
 
-    public function getStringValue(mixed $value, string $serializedValue): null|string
+    public function getStringValue(mixed $value): null|string
     {
         return null;
     }