+++ src/DataType/ScalarHandler.php
@@ -49,4 +49,9 @@ abstract class ScalarHandler implements HandlerInterface
 
         return $serializedValue;
     }
+
+    public function isIdempotent(): bool
+    {
+        return true;
+    }
 }