+++ src/DataType/SerializeHandler.php
@@ -27,13 +27,26 @@ class SerializeHandler implements HandlerInterface
         return app('encrypter')->decrypt($serializedValue, true);
     }
 
-    public function getNumericValue(mixed $value, string $serializedValue): null|int|float
+    public function getNumericValue(mixed $value): null|int|float
     {
         return null;
     }
 
-    public function getStringValue(mixed $value, string $serializedValue): null|string
+    public function getStringValue(mixed $value): null|string
     {
-        return null;
+        if (!config('metable.indexComplexDataTypes', false)) {
+            return null;
+        }
+
+        return substr(
+            serialize($value),
+            0,
+            config('metable.stringValueIndexLength', 255)
+        );
+    }
+
+    public function isIdempotent(): bool
+    {
+        return false;
     }
 }