+++ tests/Integration/MetableTest.php
@@ -369,9 +369,11 @@ class MetableTest extends TestCase
         $metable = $this->createMetable();
         $metable->setMeta('foo', 123);
 
-        $result = SampleMetable::whereMetaNumeric('foo', '>', 4)->first();
+        $result = SampleMetable::whereMetaNumeric('foo', '>', 4)->get();
+        $result2 = SampleMetable::whereMetaNumeric('foo', '<', 4)->get();
 
-        $this->assertEquals($metable->getKey(), $result->getKey());
+        $this->assertEquals([$metable->getKey()], $result->modelKeys());
+        $this->assertEquals([], $result2->modelKeys());
     }
 
     public function test_it_can_be_queried_by_in_array(): void
@@ -380,11 +382,159 @@ class MetableTest extends TestCase
         $metable = $this->createMetable();
         $metable->setMeta('foo', 'bar');
 
-        $result1 = SampleMetable::whereMetaIn('foo', ['baz', 'bar'])->first();
-        $result2 = SampleMetable::whereMetaIn('foo', ['baz', 'bat'])->first();
+        $result1 = SampleMetable::whereMetaIn('foo', ['baz', 'bar'])->get();
+        $result2 = SampleMetable::whereMetaIn('foo', ['baz', 'bat'])->get();
 
-        $this->assertEquals($metable->getKey(), $result1->getKey());
-        $this->assertNull($result2);
+        $this->assertEquals([$metable->getKey()], $result1->modelKeys());
+        $this->assertEquals([], $result2->modelKeys());
+    }
+
+
+    public function test_it_can_be_queried_by_not_in_array(): void
+    {
+        $this->useDatabase();
+        $metable = $this->createMetable();
+        $metable->setMeta('foo', 'bar');
+
+        $result1 = SampleMetable::whereMetaNotIn('foo', ['baz', 'bar'])->get();
+        $result2 = SampleMetable::whereMetaNotIn('foo', ['baz', 'bat'])->get();
+
+        $this->assertEquals([], $result1->modelKeys());
+        $this->assertEquals([$metable->getKey()], $result2->modelKeys());
+    }
+
+    public function test_it_can_be_queried_by_in_array_numeric(): void
+    {
+        $this->useDatabase();
+        $metable = $this->createMetable();
+        $metable->setMeta('foo', 1.1);
+
+        $result1 = SampleMetable::whereMetaInNumeric('foo', [1.1, 2.2])->get();
+        $result2 = SampleMetable::whereMetaInNumeric('foo', [1, 2])->get();
+
+        $this->assertEquals([$metable->getKey()], $result1->modelKeys());
+        $this->assertEquals([], $result2->modelKeys());
+    }
+
+    public function test_it_can_be_queried_by_not_in_array_numeric(): void
+    {
+        $this->useDatabase();
+        $metable = $this->createMetable();
+        $metable->setMeta('foo', 1.1);
+
+        $result1 = SampleMetable::whereMetaNotInNumeric('foo', [1.1, 2.2])->get();
+        $result2 = SampleMetable::whereMetaNotInNumeric('foo', [1, 2])->get();
+
+        $this->assertEquals([], $result1->modelKeys());
+        $this->assertEquals([$metable->getKey()], $result2->modelKeys());
+    }
+
+    public function test_it_can_be_queried_by_meta_between(): void
+    {
+        $this->useDatabase();
+        $metable = $this->createMetable();
+        $metable->setMeta('foo', 'c');
+
+        $result1 = SampleMetable::whereMetaBetween(
+            'foo',
+            'a',
+            'd'
+        )->get();
+        $result2 = SampleMetable::whereMetaBetween(
+            'foo',
+            'd',
+            'z'
+        )->get();
+
+        $this->assertEquals([$metable->getKey()], $result1->modelKeys());
+        $this->assertEquals([], $result2->modelKeys());
+    }
+
+    public function test_it_can_be_queried_by_meta_between_numeric(): void
+    {
+        $this->useDatabase();
+        $metable = $this->createMetable();
+        $date = Carbon::now();
+        $metable->setMeta('foo', $date);
+
+        $result1 = SampleMetable::whereMetaBetweenNumeric(
+            'foo',
+            $date->clone()->subDay(),
+            $date->clone()->addDay()
+        )->get();
+        $result2 = SampleMetable::whereMetaBetweenNumeric(
+            'foo',
+            $date->subDays(2),
+            $date->subDay()
+        )->get();
+
+        $this->assertEquals([$metable->getKey()], $result1->modelKeys());
+        $this->assertEquals([], $result2->modelKeys());
+    }
+
+    public function test_it_can_be_queried_by_meta_not_between_numeric(): void
+    {
+        $this->useDatabase();
+        $metable = $this->createMetable();
+        $date = Carbon::now();
+        $metable->setMeta('foo', $date);
+
+        $result1 = SampleMetable::whereMetaNotBetweenNumeric(
+            'foo',
+            $date->clone()->subDay(),
+            $date->clone()->addDay()
+        )->get();
+        $result2 = SampleMetable::whereMetaNotBetweenNumeric(
+            'foo',
+            $date->subDays(2),
+            $date->subDay()
+        )->get();
+
+        $this->assertEquals([], $result1->modelKeys());
+        $this->assertEquals([$metable->getKey()], $result2->modelKeys());
+    }
+
+    public function test_it_can_be_queried_by_null(): void
+    {
+        $this->useDatabase();
+        $metable1 = $this->createMetable();
+        $metable1->setMeta('foo', null);
+
+        $metable2 = $this->createMetable();
+        $metable2->setMeta('foo', 1);
+
+        $result1 = SampleMetable::whereMetaIsNull('foo')->get();
+
+        $this->assertEquals([$metable1->getKey()], $result1->modelKeys());
+    }
+
+    public function test_it_can_be_queried_by_model(): void
+    {
+        $this->useDatabase();
+        $metable1 = $this->createMetable();
+        $metable2 = $this->createMetable();
+        $metable1->setMeta('foo', $metable2);
+        $metable2->setMeta('foo', $metable1);
+
+        $result1 = SampleMetable::whereMetaIsModel('foo', $metable1)->get();
+        $result2 = SampleMetable::whereMetaIsModel('foo', $metable2)->get();
+        $result3 = SampleMetable::whereMetaIsModel(
+            'foo',
+            SampleMetable::class,
+            $metable1->getKey()
+        )->get();
+        $result4 = SampleMetable::whereMetaIsModel(
+            'foo',
+            SampleMetable::class,
+            $metable2->getKey()
+        )->get();
+        $result5 = SampleMetable::whereMetaIsModel('foo', SampleMetable::class)->get();
+
+        $this->assertEquals([$metable2->getKey()], $result1->modelKeys());
+        $this->assertEquals([$metable1->getKey()], $result2->modelKeys());
+        $this->assertEquals([$metable2->getKey()], $result3->modelKeys());
+        $this->assertEquals([$metable1->getKey()], $result4->modelKeys());
+        $this->assertEquals([$metable1->getKey(), $metable2->getKey()], $result5->modelKeys());
     }
 
     public function test_it_can_order_query_by_meta_value(): void
@@ -400,8 +550,8 @@ class MetableTest extends TestCase
         $results1 = SampleMetable::orderByMeta('foo', 'asc')->get();
         $results2 = SampleMetable::orderByMeta('foo', 'desc')->get();
 
-        $this->assertEquals([3, 1, 2], $results1->pluck('id')->toArray());
-        $this->assertEquals([2, 1, 3], $results2->pluck('id')->toArray());
+        $this->assertEquals([3, 1, 2], $results1->modelKeys());
+        $this->assertEquals([2, 1, 3], $results2->modelKeys());
     }
 
     public function test_it_can_order_query_by_meta_value_strict(): void
@@ -417,8 +567,8 @@ class MetableTest extends TestCase
         $results1 = SampleMetable::orderByMeta('foo', 'asc', true)->get();
         $results2 = SampleMetable::orderByMeta('foo', 'desc', true)->get();
 
-        $this->assertEquals([3, 1], $results1->pluck('id')->toArray());
-        $this->assertEquals([1, 3], $results2->pluck('id')->toArray());
+        $this->assertEquals([3, 1], $results1->modelKeys());
+        $this->assertEquals([1, 3], $results2->modelKeys());
     }
 
     public function test_it_can_order_query_by_numeric_meta_value(): void
@@ -434,8 +584,8 @@ class MetableTest extends TestCase
         $results1 = SampleMetable::orderByMetaNumeric('foo', 'asc')->get();
         $results2 = SampleMetable::orderByMetaNumeric('foo', 'desc')->get();
 
-        $this->assertEquals([2, 3, 1], $results1->pluck('id')->toArray());
-        $this->assertEquals([1, 3, 2], $results2->pluck('id')->toArray());
+        $this->assertEquals([2, 3, 1], $results1->modelKeys());
+        $this->assertEquals([1, 3, 2], $results2->modelKeys());
     }
 
     public function test_it_can_order_query_by_numeric_meta_value_strict(): void
@@ -451,8 +601,8 @@ class MetableTest extends TestCase
         $results1 = SampleMetable::orderByMetaNumeric('foo', 'asc', true)->get();
         $results2 = SampleMetable::orderByMetaNumeric('foo', 'desc', true)->get();
 
-        $this->assertEquals([3, 1], $results1->pluck('id')->toArray());
-        $this->assertEquals([1, 3], $results2->pluck('id')->toArray());
+        $this->assertEquals([3, 1], $results1->modelKeys());
+        $this->assertEquals([1, 3], $results2->modelKeys());
     }
 
     public function test_set_relation_updates_index(): void